move plotters to viz
